refactor(applications): extract draft persistence into useFormDraftPersistence hook
Moves the sessionStorage debounced persist/clear/read/flush machinery and the
hasMeaningfulFormValue helpers out of ApplicationForm into a dedicated hook,
shrinking the component from 511 to 455 lines and addressing react-doctor's
no-giant-component warning. The watch effect now collapses to a single
schedulePersistOrClear call, making the auth-persistence flow easier to reason
about and test independently.
